+++ src/main/kotlin/com/example/mykku/image/S3ImageUploadService.kt
@@ -1,8 +1,9 @@
 package com.example.mykku.image
 
 import com.example.mykku.config.S3Properties
-import com.example.mykku.image.exception.ImageException
+import com.example.mykku.image.dto.FanNoteImagesUploadResult
 import com.example.mykku.image.dto.ImageUploadResult
+import com.example.mykku.image.exception.ImageException
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
@@ -11,6 +12,7 @@ import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
 import java.awt.image.BufferedImage
 import java.io.ByteArrayInputStream
+import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 import java.util.*
@@ -27,6 +29,8 @@ class S3ImageUploadService(
         private const val MAX_FILE_SIZE = 10 * 1024 * 1024 // 10MB
         private val FILENAME_TS_FORMATTER: DateTimeFormatter =
             DateTimeFormatter.ofPattern("yyyyMMddHHmmss")
+        private val DATE_FORMATTER: DateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
+        private val FILENAME_DATE_FORMATTER: DateTimeFormatter = DateTimeFormatter.ofPattern("yyyyMMdd")
     }
 
     override fun uploadImages(images: List<MultipartFile>): List<ImageUploadResult> {
@@ -112,5 +116,92 @@ class S3ImageUploadService(
         } catch (e: Exception) {
             throw ImageException.imageSizeExtractionFailed()
         }
+    }
+
+    /**
+     * 팬노트 이미지 업로드 (커버 + 페이지)
+     * @param coverImage 커버 이미지 (optional)
+     * @param pageImages 페이지 이미지 리스트 (optional)
+     * @return 업로드된 이미지 URL 정보
+     */
+    override fun uploadFanNoteImages(
+        coverImage: MultipartFile?,
+        pageImages: List<MultipartFile>?
+    ): FanNoteImagesUploadResult {
+        val folderName = createFanNoteFolder()
+
+        val coverUrl = coverImage?.takeIf { !it.isEmpty }?.let {
+            uploadFanNoteImageToS3(it, folderName, 0)
+        }
+
+        val pageUrls = pageImages?.filterNot { it.isEmpty }?.mapIndexed { index, file ->
+            uploadFanNoteImageToS3(file, folderName, index + 1)
+        } ?: emptyList()
+
+        return FanNoteImagesUploadResult(coverUrl, pageUrls)
+    }
+
+    /**
+     * 팬노트별 고유 폴더명 생성
+     * 형식: yyyy-MM-dd-{UUID 앞 8자리}
+     */
+    private fun createFanNoteFolder(): String {
+        val date = LocalDate.now().format(DATE_FORMATTER)
+        val uuid = UUID.randomUUID().toString().replace("-", "").substring(0, 8)
+        return "$date-$uuid"
+    }
+
+    /**
+     * 팬노트 이미지를 S3에 업로드
+     * @param image 업로드할 이미지
+     * @param folderName 팬노트 폴더명
+     * @param pageNumber 페이지 번호 (0=커버, 1~=페이지)
+     * @return 업로드된 이미지 URL
+     */
+    private fun uploadFanNoteImageToS3(
+        image: MultipartFile,
+        folderName: String,
+        pageNumber: Int
+    ): String {
+        validateImage(image)
+
+        val imageBytes = image.bytes
+        val fileName = generateFanNoteFileName(image.originalFilename, pageNumber)
+        val key = "fan-note/$folderName/$fileName"
+
+        val extensionForContentType = getFileExtension(image.originalFilename)
+        val contentType = image.contentType ?: when (extensionForContentType) {
+            "jpg", "jpeg" -> "image/jpeg"
+            "png" -> "image/png"
+            "gif" -> "image/gif"
+            "webp" -> "image/webp"
+            else -> "application/octet-stream"
+        }
+
+        val putObjectRequest = PutObjectRequest.builder()
+            .bucket(s3Properties.bucketName)
+            .key(key)
+            .contentType(contentType)
+            .contentLength(imageBytes.size.toLong())
+            .build()
+
+        s3Client.putObject(
+            putObjectRequest,
+            RequestBody.fromBytes(imageBytes)
+        )
+
+        return s3Client.utilities()
+            .getUrl { it.bucket(s3Properties.bucketName).key(key) }
+            .toString()
+    }
+
+    /**
+     * 팬노트 파일명 생성
+     * 형식: yyyyMMdd-{pageNumber}.{extension}
+     */
+    private fun generateFanNoteFileName(originalFilename: String?, pageNumber: Int): String {
+        val extension = getFileExtension(originalFilename)
+        val date = LocalDate.now().format(FILENAME_DATE_FORMATTER)
+        return "$date-$pageNumber.$extension"
     }
 }
