+++ src/main/kotlin/com/example/mykku/image/dto/FanNoteImagesUploadResult.kt
@@ -1,0 +1,6 @@
+package com.example.mykku.image.dto
+
+data class FanNoteImagesUploadResult(
+    val coverImageUrl: String?,
+    val pageImageUrls: List<String>
+)
